Changes schematic so the board is never powered by the LED strip
The idea is that since this is a tester, it will provide the power to
the LED strip. We add a schottky diode to protect the board just in case
it is connected to an already powered strip.

Since we'll always power the board from a usb-c connector, we don't need
reverse polarity protection anymore.