Adds simple firmware for testing LED strips based on FAB_LED
This is basically copied from the examples in FAB_LED[1]. FAB_LED is a
very compact library for controlling ws2812 (and other) LEDs and lets
you control a very large number of pixels with very little memory usage.

To work on an attiny85 you need to set the microcontroller to use
internal PLL for a 16MHz clock.

[1]: https://github.com/sonyhome/FAB_LED